requirements.txt
websockets>=12.0
fastapi>=0.111.0
uvicorn>=0.29.0
claude-agent-sdk>=0.0.14
anthropic>=0.52.0
python-dotenv>=1.0.0
python-socketio>=5.11.0
httpx>=0.27.0
